updated POST and PUT endpoints

+++ routes/tags.js
@@ -53,8 +53,9 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
     const { name } = req.body;
+    const userId = req.user.id;
 
-    const newTag = { name };
+    const newTag = { name, userId };
 
     /***** Never trust users - validate input *****/
     if (!name) {
@@ -80,6 +81,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
     const { id } = req.params;
     const { name } = req.body;
+    const userId = req.user.id;
 
     /***** Never trust users - validate input *****/
     if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -94,9 +96,9 @@ router.put('/:id', (req, res, next) => {
         return next(err);
     }
 
-    const updateTag = { name };
+    const updateTag = { name, userId };
 
-    Tag.findByIdAndUpdate(id, updateTag, { new: true })
+    Tag.find(id, updateTag, { new: true })
         .then(result => {
             if (result) {
                 res.json(result);
@@ -116,6 +118,7 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
     const { id } = req.params;
+    const userId = req.user.id;
 
     /***** Never trust users - validate input *****/
     if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -124,7 +127,7 @@ router.delete('/:id', (req, res, next) => {
         return next(err);
     }
 
-    const tagRemovePromise = Tag.findByIdAndRemove(id);
+    const tagRemovePromise = Tag.findByIdAndRemove({_id: id, userId});
 
     const noteUpdatePromise = Note.updateMany(
         { tags: id, },
